itp_iterative_search_lm/requirements.txt
ctransformers[cuda] >= 0.2.27
jaxtyping >= 0.2.25
# datasets >= 2.17.0
torch >= 2.1.2
transformers >= 4.37.2
vllm >= 0.3.1

hydra-core >= 1.3.2
pandas >= 2.2.0

ipdb >= 0.13.13
ipython >= 8.21.0
rich >= 13.7.0
# notebook >= 7.0.8

toolz >= 0.12.1
